renterd: move stores package out of internal
