+++ cmd/renterd/main.go
@@ -20,8 +20,8 @@ import (
 	"go.sia.tech/renterd/build"
 	"go.sia.tech/renterd/bus"
 	"go.sia.tech/renterd/internal/node"
-	"go.sia.tech/renterd/internal/stores"
 	"go.sia.tech/renterd/internal/tracing"
+	"go.sia.tech/renterd/stores"
 	"go.sia.tech/renterd/wallet"
 	"go.sia.tech/renterd/worker"
 	"golang.org/x/term"
