+++ stores/autopilot.go
