+++ stores/hostdb.go
