+++ stores/metadata_test.go
