+++ stores/settingsdb_test.go
